refactor(team): 팀 도메인 패키지 이동
