+++ backend/src/main/java/com/morak/back/poll/application/PollService.java
@@ -2,9 +2,6 @@ package com.morak.back.poll.application;
 
 import com.morak.back.auth.domain.Member;
 import com.morak.back.auth.domain.MemberRepository;
-import com.morak.back.auth.domain.Team;
-import com.morak.back.auth.domain.TeamMemberRepository;
-import com.morak.back.auth.domain.TeamRepository;
 import com.morak.back.auth.exception.MemberNotFoundException;
 import com.morak.back.auth.exception.TeamNotFoundException;
 import com.morak.back.core.util.CodeGenerator;
@@ -20,6 +17,9 @@ import com.morak.back.poll.ui.dto.PollItemRequest;
 import com.morak.back.poll.ui.dto.PollItemResponse;
 import com.morak.back.poll.ui.dto.PollItemResultResponse;
 import com.morak.back.poll.ui.dto.PollResponse;
+import com.morak.back.team.domain.Team;
+import com.morak.back.team.domain.TeamMemberRepository;
+import com.morak.back.team.domain.TeamRepository;
 import com.morak.back.team.exception.MismatchedTeamException;
 import java.util.List;
 import java.util.Map;
