+++ backend/src/main/java/com/morak/back/poll/domain/BaseEntity.java
@@ -1,15 +1,12 @@
 package com.morak.back.poll.domain;
 
 import java.time.LocalDateTime;
-
 import javax.persistence.EntityListeners;
 import javax.persistence.MappedSuperclass;
-
+import lombok.Getter;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import lombok.Getter;
 
 @Getter
 @MappedSuperclass
