+++ backend/src/main/java/com/morak/back/poll/domain/Poll.java
@@ -1,8 +1,8 @@
 package com.morak.back.poll.domain;
 
 import com.morak.back.auth.domain.Member;
-import com.morak.back.auth.domain.Team;
 import com.morak.back.core.exception.InvalidRequestException;
+import com.morak.back.team.domain.Team;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
