+++ backend/src/main/java/com/morak/back/poll/domain/PollRepository.java
@@ -2,7 +2,6 @@ package com.morak.back.poll.domain;
 
 import java.util.List;
 import java.util.Optional;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PollRepository extends JpaRepository<Poll, Long> {
