+++ backend/src/main/java/com/morak/back/team/application/TeamService.java
@@ -2,17 +2,17 @@ package com.morak.back.team.application;
 
 import com.morak.back.auth.domain.Member;
 import com.morak.back.auth.domain.MemberRepository;
-import com.morak.back.auth.domain.Team;
-import com.morak.back.auth.domain.TeamMember;
-import com.morak.back.auth.domain.TeamMemberRepository;
-import com.morak.back.auth.domain.TeamRepository;
 import com.morak.back.auth.exception.MemberNotFoundException;
 import com.morak.back.auth.exception.TeamNotFoundException;
 import com.morak.back.auth.ui.dto.MemberResponse;
-import com.morak.back.core.util.CodeGenerator;
 import com.morak.back.core.exception.ResourceNotFoundException;
+import com.morak.back.core.util.CodeGenerator;
+import com.morak.back.team.domain.Team;
 import com.morak.back.team.domain.TeamInvitation;
 import com.morak.back.team.domain.TeamInvitationRepository;
+import com.morak.back.team.domain.TeamMember;
+import com.morak.back.team.domain.TeamMemberRepository;
+import com.morak.back.team.domain.TeamRepository;
 import com.morak.back.team.exception.AlreadyJoinedTeamException;
 import com.morak.back.team.exception.ExpiredInvitationException;
 import com.morak.back.team.exception.MismatchedTeamException;
