+++ backend/src/main/java/com/morak/back/team/domain/Team.java
@@ -1,4 +1,4 @@
-package com.morak.back.auth.domain;
+package com.morak.back.team.domain;
 
 import com.morak.back.poll.domain.BaseEntity;
 import javax.persistence.Entity;
