+++ backend/src/main/java/com/morak/back/team/domain/TeamMember.java
@@ -1,5 +1,6 @@
-package com.morak.back.auth.domain;
+package com.morak.back.team.domain;
 
+import com.morak.back.auth.domain.Member;
 import com.morak.back.poll.domain.BaseEntity;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
