+++ backend/src/main/java/com/morak/back/team/domain/TeamMemberRepository.java
@@ -1,4 +1,4 @@
-package com.morak.back.auth.domain;
+package com.morak.back.team.domain;
 
 import java.util.List;
 import java.util.Optional;
