+++ backend/src/main/java/com/morak/back/team/domain/TeamRepository.java
@@ -1,4 +1,4 @@
-package com.morak.back.auth.domain;
+package com.morak.back.team.domain;
 
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
