+++ backend/src/main/java/com/morak/back/team/ui/dto/TeamCreateRequest.java
@@ -1,7 +1,6 @@
 package com.morak.back.team.ui.dto;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
-
 import javax.validation.constraints.NotBlank;
 import lombok.Getter;
