+++ backend/src/test/java/com/morak/back/auth/domain/TeamRepositoryTest.java
@@ -1,7 +1,9 @@
 package com.morak.back.auth.domain;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
+import com.morak.back.team.domain.Team;
+import com.morak.back.team.domain.TeamRepository;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
