+++ backend/src/test/java/com/morak/back/poll/acceptance/PollAcceptanceTest.java
@@ -15,7 +15,6 @@ import io.restassured.response.Response;
 import java.time.LocalDateTime;
 import java.util.List;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
