+++ backend/src/test/java/com/morak/back/poll/application/PollServiceTest.java
@@ -12,20 +12,20 @@ import static org.mockito.MockitoAnnotations.openMocks;
 
 import com.morak.back.auth.domain.Member;
 import com.morak.back.auth.domain.MemberRepository;
-import com.morak.back.auth.domain.Team;
-import com.morak.back.auth.domain.TeamMemberRepository;
-import com.morak.back.auth.domain.TeamRepository;
+import com.morak.back.core.exception.InvalidRequestException;
 import com.morak.back.poll.domain.Poll;
 import com.morak.back.poll.domain.PollItem;
 import com.morak.back.poll.domain.PollItemRepository;
 import com.morak.back.poll.domain.PollRepository;
 import com.morak.back.poll.domain.PollResult;
-import com.morak.back.core.exception.InvalidRequestException;
 import com.morak.back.poll.ui.dto.PollCreateRequest;
 import com.morak.back.poll.ui.dto.PollItemRequest;
 import com.morak.back.poll.ui.dto.PollItemResponse;
 import com.morak.back.poll.ui.dto.PollItemResultResponse;
 import com.morak.back.poll.ui.dto.PollResponse;
+import com.morak.back.team.domain.Team;
+import com.morak.back.team.domain.TeamMemberRepository;
+import com.morak.back.team.domain.TeamRepository;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Arrays;
