+++ backend/src/test/java/com/morak/back/poll/domain/PollRepositoryTest.java
@@ -3,9 +3,9 @@ package com.morak.back.poll.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.morak.back.auth.domain.Member;
-import com.morak.back.auth.domain.Team;
 import com.morak.back.poll.support.DomainSupplier;
 import com.morak.back.support.RepositoryTest;
+import com.morak.back.team.domain.Team;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
