+++ backend/src/test/java/com/morak/back/poll/domain/PollStatusTest.java
@@ -1,6 +1,7 @@
 package com.morak.back.poll.domain;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.morak.back.core.exception.InvalidRequestException;
 import org.junit.jupiter.api.Test;
