+++ backend/src/test/java/com/morak/back/team/application/TeamServiceTest.java
@@ -11,16 +11,16 @@ import static org.mockito.Mockito.verify;
 
 import com.morak.back.auth.domain.Member;
 import com.morak.back.auth.domain.MemberRepository;
-import com.morak.back.auth.domain.Team;
-import com.morak.back.auth.domain.TeamMember;
-import com.morak.back.auth.domain.TeamMemberRepository;
-import com.morak.back.auth.domain.TeamRepository;
 import com.morak.back.auth.ui.dto.MemberResponse;
 import com.morak.back.core.exception.ResourceNotFoundException;
 import com.morak.back.team.domain.ExpiredTime;
 import com.morak.back.team.domain.InvitationCode;
+import com.morak.back.team.domain.Team;
 import com.morak.back.team.domain.TeamInvitation;
 import com.morak.back.team.domain.TeamInvitationRepository;
+import com.morak.back.team.domain.TeamMember;
+import com.morak.back.team.domain.TeamMemberRepository;
+import com.morak.back.team.domain.TeamRepository;
 import com.morak.back.team.exception.AlreadyJoinedTeamException;
 import com.morak.back.team.exception.ExpiredInvitationException;
 import com.morak.back.team.exception.MismatchedTeamException;
