+++ backend/src/test/java/com/morak/back/team/domain/TeamInvitationTest.java
@@ -1,9 +1,7 @@
 package com.morak.back.team.domain;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
-import java.time.LocalDateTime;
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class TeamInvitationTest {
